Changed: merge CompressedHistory and History into one class History.
