+++ src/model/history.js
@@ -8,11 +8,14 @@
 import CKEditorError from '../../utils/ckeditorerror.js';
 
 /**
- * `History` keeps the track of all the deltas applied to the {@link engine.model.Document document}. `History` can be
- * seen as "add-only" structure. You can read and add deltas, but can't modify them. Use this version of history to
- * retrieve applied deltas as they were, in the original form.
+ * `History` keeps the track of all the deltas applied to the {@link engine.model.Document document}. Deltas stored in
+ * `History` might get updated, split into more deltas or even removed. This is used mostly to compress history, instead
+ * of keeping all deltas in a state in which they were applied.
  *
- * @see engine.model.CompressedHistory
+ * **Note:** deltas kept in `History` should be used only to transform deltas. It's not advised to use `History` to get
+ * original delta basing on it's {@link engine.model.delta.Delta#baseVersion baseVersion}. Also, after transforming a
+ * delta by deltas from `History`, fix it's base version accordingly (set to {@link engine.model.Document#version}.
+ *
  * @memberOf engine.model
  */
 export default class History {
@@ -88,15 +91,90 @@ export default class History {
 	}
 
 	/**
-	 * Returns a delta from the history that has given {@link engine.model.delta.Delta#baseVersion}.
+	 * Returns one or more deltas from history that bases on given `baseVersion`. Most often it will be just
+	 * one delta, but if that delta got updated by multiple deltas, all of those updated deltas will be returned.
 	 *
+	 * @see engine.model.History#updateDelta
 	 * @param {Number} baseVersion Base version of the delta to retrieve.
-	 * @returns {engine.model.delta.Delta|null} Delta with given base version or null if no such delta is in history.
+	 * @returns {Array.<engine.model.delta.Delta>|null} Delta with given base version or null if no such delta is in history.
 	 */
 	getDelta( baseVersion ) {
 		let index = this._historyPoints.get( baseVersion );
 
-		return this._deltas[ index ] || null;
+		if ( index === undefined ) {
+			return null;
+		}
+
+		const deltas = [];
+
+		for ( index; index < this._deltas.length; index++ ) {
+			const delta = this._deltas[ index ];
+
+			if ( delta.baseVersion != baseVersion ) {
+				break;
+			}
+
+			deltas.push( delta );
+		}
+
+		return deltas.length === 0 ? null : deltas;
+	}
+
+	/**
+	 * Removes delta from the history. This happens i.e., when a delta is undone by another delta. Both undone delta and
+	 * undoing delta should be removed so they won't have an impact on transforming other deltas.
+	 *
+	 * **Note:** using this method does not change the state of {@link engine.model.Document model}. It just affects
+	 * the state of `History`.
+	 *
+	 * **Note:** when some deltas are removed, deltas between them should probably get updated. See
+	 * {@link engine.model.History#updateDelta}.
+	 *
+	 * **Note:** if delta with `baseVersion` got {@link engine.model.History#updateDelta updated} by multiple
+	 * deltas, all updated deltas will be removed.
+	 *
+	 * @param {Number} baseVersion Base version of a delta to be removed.
+	 */
+	removeDelta( baseVersion ) {
+		this.updateDelta( baseVersion, [] );
+	}
+
+	/**
+	 * Substitutes delta in history by one or more given deltas.
+	 *
+	 * **Note:** if delta with `baseVersion` was already updated by multiple deltas, all updated deltas will be removed
+	 * and new deltas will be inserted at their position.
+	 *
+	 * **Note:** delta marked as reversed won't get updated.
+	 *
+	 * @param {Number} baseVersion Base version of a delta to update.
+	 * @param {Iterable.<engine.model.delta.Delta>} updatedDeltas Deltas to be inserted in place of updated delta.
+	 */
+	updateDelta( baseVersion, updatedDeltas ) {
+		const deltas = this.getDelta( baseVersion );
+
+		// If there are no deltas, stop executing function as there is nothing to update.
+		if ( deltas === null ) {
+			return;
+		}
+
+		// Make sure that every updated delta has correct `baseVersion`.
+		// This is crucial for algorithms in `History` and algorithms using `History`.
+		for ( let delta of updatedDeltas ) {
+			delta.baseVersion = baseVersion;
+		}
+
+		// Put updated deltas in place of old deltas.
+		this._deltas.splice( this._getIndex( baseVersion ), deltas.length, ...updatedDeltas );
+
+		// Update history points.
+		const changeBy = updatedDeltas.length - deltas.length;
+
+		for ( let key of this._historyPoints.keys() ) {
+			if ( key > baseVersion ) {
+				this._historyPoints.set( key, this._historyPoints.get( key ) + changeBy );
+			}
+		}
 	}
 
 	/**
