+++ tests/model/history.js
@@ -68,6 +68,41 @@ describe( 'History', () => {
 		} );
 	} );
 
+	describe( 'getDelta', () => {
+		it( 'should return array with one delta with given base version', () => {
+			let delta = getDelta( 0 );
+			history.addDelta( delta );
+
+			const historyDelta = history.getDelta( 0 );
+			expect( historyDelta ).to.deep.equal( [ delta ] );
+		} );
+
+		it( 'should return array with all updated deltas of delta with given base version', () => {
+			let delta = getDelta( 0 );
+			history.addDelta( delta );
+
+			let deltas = getDeltaSet();
+			history.updateDelta( 0, deltas );
+
+			const historyDelta = history.getDelta( 0 );
+			expect( historyDelta ).to.deep.equal( deltas );
+		} );
+
+		it( 'should return null if delta has not been found in history', () => {
+			expect( history.getDelta( -1 ) ).to.be.null;
+			expect( history.getDelta( 2 ) ).to.be.null;
+			expect( history.getDelta( 20 ) ).to.be.null;
+		} );
+
+		it( 'should return null if delta has been removed by removeDelta', () => {
+			let delta = getDelta( 0 );
+			history.addDelta( delta );
+			history.removeDelta( 0 );
+
+			expect( history.getDelta( 0 ) ).to.be.null;
+		} );
+	} );
+
 	describe( 'getDeltas', () => {
 		let deltas;
 
@@ -101,23 +136,79 @@ describe( 'History', () => {
 		} );
 	} );
 
-	describe( 'getDelta', () => {
-		beforeEach( () => {
+	describe( 'updateDelta', () => {
+		it( 'should substitute delta from history by given deltas', () => {
+			history.addDelta( getDelta( 0 ) );
+
+			const deltas = getDeltaSet();
+			history.updateDelta( 0, deltas );
+
+			const historyDeltas = Array.from( history.getDeltas() );
+			expect( historyDeltas ).to.deep.equal( deltas );
+		} );
+
+		it( 'should substitute all updated deltas by new deltas', () => {
+			history.addDelta( getDelta( 0 ) );
+
+			// Change original single delta to three deltas generated by `getDeltaSet`.
+			// All those deltas should now be seen under base version 0.
+			history.updateDelta( 0, getDeltaSet() );
+
+			const deltas = getDeltaSet();
+			// Change all three deltas from base version 0 to new set of deltas.
+			history.updateDelta( 0, deltas );
+
+			const historyDeltas = Array.from( history.getDeltas() );
+			expect( historyDeltas ).to.deep.equal( deltas );
+		} );
+
+		it( 'should do nothing if deltas for given base version has not been found in history', () => {
+			history.addDelta( getDelta( 0 ) );
+			history.removeDelta( 0 );
+
+			const deltas = getDeltaSet();
+
+			history.updateDelta( 0, deltas );
+
+			expect( Array.from( history.getDeltas() ).length ).to.equal( 0 );
+		} );
+	} );
+
+	describe( 'removeDelta', () => {
+		it( 'should remove deltas that do not have graveyard related operations', () => {
 			for ( let delta of getDeltaSet() ) {
 				history.addDelta( delta );
 			}
+
+			history.removeDelta( 3 );
+
+			const deltas = Array.from( history.getDeltas() );
+			expect( deltas.length ).to.equal( 2 );
 		} );
 
-		it( 'should return delta from history that has given base version', () => {
-			let delta = history.getDelta( 3 );
+		it( 'should remove multiple updated deltas', () => {
+			let delta = getDelta( 0 );
+			history.addDelta( delta );
 
-			expect( delta.baseVersion ).to.equal( 3 );
+			let updatedDeltas = getDeltaSet( 0 );
+
+			history.updateDelta( 0, updatedDeltas );
+			history.removeDelta( 0 );
+
+			const deltas = Array.from( history.getDeltas() );
+			expect( deltas.length ).to.equal( 0 );
 		} );
 
-		it( 'should return null if delta has not been found in history', () => {
-			expect( history.getDelta( -1 ) ).to.be.null;
-			expect( history.getDelta( 2 ) ).to.be.null;
-			expect( history.getDelta( 20 ) ).to.be.null;
+		it( 'should do nothing if deltas for given base version has not been found in history', () => {
+			const deltas = getDeltaSet();
+
+			for ( let delta of deltas ) {
+				history.addDelta( delta );
+			}
+
+			history.removeDelta( 12 );
+
+			expect( Array.from( history.getDeltas() ) ).to.deep.equal( deltas );
 		} );
 	} );
 } );
